--- a/src/TrixiParticles.jl
+++ b/src/TrixiParticles.jl
@@ -85,7 +85,7 @@ export SourceTermDamping
 export ShepardKernelCorrection, KernelCorrection, AkinciFreeSurfaceCorrection,
        GradientCorrection, BlendedGradientCorrection, MixedKernelGradientCorrection
 export nparticles
-export kinetic_energy, total_mass, max_pressure, min_pressure, avg_pressure,
+export available_data, kinetic_energy, total_mass, max_pressure, min_pressure, avg_pressure,
        max_density, min_density, avg_density
 export interpolate_line, interpolate_point, interpolate_plane_3d, interpolate_plane_2d,
        interpolate_plane_2d_vtk