--- a/src/general/density_calculators.jl
+++ b/src/general/density_calculators.jl
@@ -23,31 +23,6 @@ difference of the coordinates, ``v_{ab} = v_a - v_b`` of the velocities of parti
 """
 struct ContinuityDensity end
 
-@propagate_inbounds function particle_density(v, system, particle)
-    particle_density(v, system.density_calculator, system, particle)
-end
-
-@propagate_inbounds function particle_density(v, ::SummationDensity, system, particle)
-    return system.cache.density[particle]
-end
-
-@propagate_inbounds function particle_density(v, ::ContinuityDensity, system, particle)
-    return v[end, particle]
-end
-
-# WARNING!
-# These functions are intended to be used internally to set the density
-# of newly activated particles in a callback.
-# DO NOT use outside a callback. OrdinaryDiffEq does not allow changing `v` and `u`
-# outside of callbacks.
-@inline set_particle_density!(v, system, ::SummationDensity, particle, density) = v
-
-@inline function set_particle_density!(v, system, ::ContinuityDensity, particle, density)
-    v[end, particle] = density
-
-    return v
-end
-
 function summation_density!(system, semi, u, u_ode, density;
                             particles=each_moving_particle(system))
     set_zero!(density)