Closes #424 and #547.

The fact that `semi` is now passed to the interpolation function by the postprocess cb instead of being a global variable also makes the interpolation 4x faster:
main:
```
apply postprocess cb                10     256s   88.5%   25.6s   89.5GiB   99.8%  8.95GiB
  ~apply postprocess cb~            10     256s   88.4%   25.6s   89.5GiB   99.8%  8.95GiB
  update systems                    10    159ms    0.1%  15.9ms   56.2KiB    0.0%  5.62KiB
```
this PR:
```
apply postprocess cb                10    64.4s   62.3%   6.44s   89.3GiB   99.7%  8.93GiB
  ~apply postprocess cb~            10    64.2s   62.1%   6.42s   89.3GiB   99.7%  8.93GiB
  update systems                    10    167ms    0.2%  16.7ms   56.2KiB    0.0%  5.62KiB
```

I also made the internal API for particle quantities more consistent by renaming `particle_density` and `particle_pressure` to `current_density` and `current_pressure`, similar to `current_coords`.